+++ src/variables/aboutConsts.js
@@ -1,8 +1,8 @@
 const GLOBAL_NAME = "BubbleOS";
 const AUTHOR = "Arnav Thorat";
 
-const VERSION = "0.6.3";
-const BUILD = "63";
+const VERSION = "0.6.6";
+const BUILD = "66";
 const IN_BETA = true;
 
 module.exports = { GLOBAL_NAME, AUTHOR, VERSION, BUILD, IN_BETA };